+++ source/backend/database/models/completedRecipesModel.js
@@ -1,0 +1,16 @@
+const db = require('../dbConfig');
+
+/**
+ * Searches all completed recipes by userId
+ * @param userId
+ * @returns {Promise<awaited Knex.QueryBuilder<TRecord, TResult>>}
+ */
+async function getCompletedRecipes(userId){
+  const completedRecipes = await db('completedRecipes')
+      .innerJoin('recipes', 'completedRecipes.recipeId', 'recipes.recipeId')
+      .where('completedRecipes.userId', userId)
+      .select('recipes.*');
+  return completedRecipes;
+}
+
+module.exports = { getCompletedRecipes };
