+++ source/backend/routes/users.js
@@ -1,5 +1,7 @@
 const express = require('express');
 const usersModel = require('../database/models/usersModel');
+const completedRecipesModel = require('../database/models/completedRecipesModel');
+const savedRecipesModel = require('../database/models/savedRecipesModel');
 
 const router = express.Router();
 
@@ -7,10 +9,10 @@ const router = express.Router();
 router.get('/completedRecipes', async (req, res) => {
   try {
     const { userId } = req.userInfo;
-    const completedRecipesList = await usersModel.getCompletedRecipes(userId);
-    res.send(completedRecipesList);
+    const completedRecipesList = await completedRecipesModel.getCompletedRecipes(userId);
+    return res.status(200).json(completedRecipesList);
   } catch (err) {
-    console.log(err);
+    console.error(err);
     return res.status(503).json({
       message: 'Failed to get completed recipes',
       err
@@ -22,10 +24,10 @@ router.get('/completedRecipes', async (req, res) => {
 router.get('/savedRecipes', async (req, res) => {
   try {
     const { userId }  = req.userInfo;
-    const savedRecipesList = await usersModel.getSavedRecipes(userId);
+    const savedRecipesList = await savedRecipesModel.getSavedRecipes(userId);
     return res.status(200).json(savedRecipesList);
   } catch (err) {
-    console.log(err);
+    console.error(err);
     return res.status(503).json({
       message: 'Failed to get saved recipes',
       err
